// src/seed/types.ts
export interface SeedData {
  organizations: Array<typeof organizations.$inferInsert>;
  actors: Array<typeof actors.$inferInsert>;
  visitors: Array<typeof visitors.$inferInsert>;
  wip: Array<typeof wip.$inferInsert>;
  auditEvents: Array<typeof auditEvents.$inferInsert>;
  serviceOfferings: Array<typeof serviceOfferings.$inferInsert>;
  drafts: Array<typeof drafts.$inferInsert>;
  modelCalls: Array<typeof modelCalls.$inferInsert>;
  draftTurns: Array<typeof draftTurns.$inferInsert>;
  serviceCandidates: Array<typeof serviceCandidates.$inferInsert>;
  requests: Array<typeof requests.$inferInsert>;
  taskCompletions: Array<typeof taskCompletions.$inferInsert>;
  requestContentRevisions: Array<typeof requestContentRevisions.$inferInsert>;
  clarificationRequests: Array<typeof clarificationRequests.$inferInsert>;
  deliveryHandoffs: Array<typeof deliveryHandoffs.$inferInsert>;
  requestResolutions: Array<typeof requestResolutions.$inferInsert>;
  inventorySources: Array<typeof inventorySources.$inferInsert>;
  inventorySyncRuns: Array<typeof inventorySyncRuns.$inferInsert>;
  inventorySourceRecords: Array<typeof inventorySourceRecords.$inferInsert>;
  inventoryAliases: Array<typeof inventoryAliases.$inferInsert>;
  inventoryConflicts: Array<typeof inventoryConflicts.$inferInsert>;
  inventoryConflictMembers: Array<typeof inventoryConflictMembers.$inferInsert>;
  catalogItems: Array<typeof catalogItems.$inferInsert>;
  catalogFieldDecisions: Array<typeof catalogFieldDecisions.$inferInsert>;
  catalogItemSources: Array<typeof catalogItemSources.$inferInsert>;
  reviewTasks: Array<typeof reviewTasks.$inferInsert>;
  assetAssessments: Array<typeof assetAssessments.$inferInsert>;
  assetCandidates: Array<typeof assetCandidates.$inferInsert>;
  assetCandidateDecisions: Array<typeof assetCandidateDecisions.$inferInsert>;
  policyRules: Array<typeof policyRules.$inferInsert>;
  riskAssessments: Array<typeof riskAssessments.$inferInsert>;
  riskFindings: Array<typeof riskFindings.$inferInsert>;
  riskFindingDecisions: Array<typeof riskFindingDecisions.$inferInsert>;
  riceScores: Array<typeof riceScores.$inferInsert>;
  workSystems: Array<typeof workSystems.$inferInsert>;
  workSyncRuns: Array<typeof workSyncRuns.$inferInsert>;
  externalWorkItems: Array<typeof externalWorkItems.$inferInsert>;
  requestWorkItemLinks: Array<typeof requestWorkItemLinks.$inferInsert>;
}